% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theil.2d.R
\name{theil.2d}
\alias{theil.2d}
\title{Function performing two-dimensional decomposition of the Theil index.}
\usage{
theil.2d(
  data,
  total,
  feature = NULL,
  sources = NULL,
  weights = NULL,
  perc = FALSE
)
}
\arguments{
\item{data}{Data frame containing income data.}

\item{total}{String specifying the name of the column containing data on
total income.}

\item{feature}{String specifying the name of the column containing
information about the feature used for inequality decomposition. If left
blank, total income is not decomposed by feature.}

\item{sources}{Vector containing strings specifying the names of the columns
with data on income sources, the sum of which must be equal to total income.
The user can specify the same value as in "total." If left blank, total
income is not decomposed by income source.}

\item{weights}{String specifying the name of the column containing population
weights.}

\item{perc}{If set to TRUE, then the function returns percentage shares of
every inequality component in overall inequality. Set to FALSE by default.}
}
\value{
Data frame containing values of components of the Theil index.
Every column represents a feature, while every row represents an income
source. Thus, every value in this data frame represents the contribution of
income inequality in i-th income source among population members
possessing j-th feature to total income inequality.
}
\description{
The function performs two-dimensional decomposition of the
Theil index according to Giammatteo (2007). That is, the index can be
decomposed by some feature that members of the studied population possess
(e.g., sex, education, age) and their income source at the same time.
}
\examples{
\dontrun{
Load the test data set.
data("us16")

No decomposition, just Theil index of total income.
result <- theil.2d(us16, "hitotal", weights = "hpopwgt")

Decomposition of income inequality by gender.
result <- theil.2d(us16, "hitotal", "sex", "hitotal", "hpopwgt")

Decomposition of income inequality by gender and income source.
result <- theil.2d(us16, "hitotal", "sex", c("hilabour", "hicapital",
"hitransfer"), "hpopwgt")
}
}
\references{
Giammatteo, M. (2007). The Bidimensional Decomposition of Inequality:
A nested Theil Approach. LIS Working papers, Article 466, 1-30.
}
