% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scv.2d.R
\name{scv.2d}
\alias{scv.2d}
\title{Function performing two-dimensional decomposition of the squared
coefficient of variation (SCV).}
\usage{
scv.2d(data, total, feature = NULL, sources = NULL, weights = NULL)
}
\arguments{
\item{data}{Data frame containing income data.}

\item{total}{String specifying the name of the column containing data on
total income.}

\item{feature}{String specifying the name of the column containing
information about the feature used for inequality decomposition. If left
blank, total income is not decomposed by feature.}

\item{sources}{Vector containing strings specifying the names of the columns
with data on income sources, the sum of which must be equal to total income.
The user can specify the same value as in "total." If left blank, total
income is not decomposed by income source.}

\item{weights}{String specifying the name of the column containing population
weights.}
}
\value{
Data frame containing values of components of SCV.
Every column represents a feature, while every row represents an income
source. Thus, every value in this data frame represents the contribution of
income inequality in i-th income source among population members
possessing j-th feature to total income inequality. These values are
calculated using the formulas suggested in Garcia-Penalosa & Orgiazzi (2013).
}
\description{
Function performing two-dimensional decomposition of the squared
coefficient of variation (SCV).
}
\examples{
\dontrun{
No decomposition, just SCV of total income.
result <- scv.2d(data, "hitotal", weights = "hpopwgt")

Decomposition of income inequality by gender.
result <- scv.2d(data, "hitotal", "sex", "hitotal", "hpopwgt")

Decomposition of income inequality by gender and income source.
result <- scv.2d(data, "hitotal", "sex", c("hilabour", "hicapital",
"hitransfer"), "hpopwgt")
}
}
\references{
Garcia-Penalosa, C., & Orgiazzi, E. (2013). Factor Components of Inequality:
A Cross-Country Study. Review of Income and Wealth, 59(4), 689-727.
}
